+++ sdk/datadog/arm-datadog/src/operations/index.ts
@@ -10,6 +10,7 @@ export * from "./marketplaceAgreements";
 export * from "./creationSupported";
 export * from "./monitors";
 export * from "./operations";
+export * from "./billingInfo";
 export * from "./tagRules";
 export * from "./singleSignOnConfigurations";
 export * from "./monitoredSubscriptions";